requirements.txt
apache-airflow==2.6.3
boto3
pandas
sqlalchemy
psycopg2-binary
python-dotenv
jsonschema
pytest
